+++ app/Models/Order.php
@@ -25,15 +25,6 @@ class Order extends Model
     ];
 
     /**
-     * The attributes that should be hidden for arrays.
-     *
-     * @var array
-     */
-    protected $hidden = [
-        'created_at', 'updated_at'
-    ];
-
-    /**
      * The attributes that should be cast to native types.
      *
      * @var array
