+++ routes/api.php
@@ -3,8 +3,7 @@
 use App\Http\Controllers\API\PizzaController;
 use App\Http\Controllers\API\UserController;
 use App\Http\Controllers\API\CartController;
-use App\Http\Controllers\OrderController;
-use Illuminate\Http\Request;
+use App\Http\Controllers\API\OrderController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -19,6 +18,19 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::group(['prefix' => 'v1'], function () {
+    Route::middleware(['jwt.auth'])->group(function () {
+        // Profile
+        Route::patch('profile', [UserController::class, 'updateProfile']);
+        Route::get('profile', [UserController::class, 'getProfile']);
+
+        // pizza
+        Route::post('pizza', [PizzaController::class, 'addPizza']);
+        Route::delete('pizza/{id}',  [PizzaController::class, 'deletePizza']);
+
+        // order
+        Route::get('order/history', [OrderController::class, 'getOrderHistory']);
+    });
+
     // Auth
     Route::post('auth/register', [UserController::class, 'register']);
     Route::post('auth/login', [UserController::class, 'login']);
@@ -39,14 +51,5 @@ Route::group(['prefix' => 'v1'], function () {
     // Order
     Route::post('order', [OrderController::class, 'createOrder']);
     Route::get('order/{id}', [OrderController::class, 'getOrderSummary']);
-
-    Route::middleware(['jwt.auth'])->group(function () {
-        // Profile
-        Route::patch('profile', [UserController::class, 'updateProfile']);
-        Route::get('profile', [UserController::class, 'getProfile']);
-
-        //pizza
-        Route::post('pizza', [PizzaController::class, 'addPizza']);
-        Route::delete('pizza/{id}',  [PizzaController::class, 'deletePizza']);
-    });
+    Route::delete('order/{id}', [OrderController::class, 'cancelOrder']);
 });
